Fixed the stats not incrementing properly

+++ Assets/Scripts/Upgrades/UpgradeStats/SpeedUpgradeStats.cs
@@ -3,7 +3,7 @@
 [CreateAssetMenu(fileName = "Upgrade Stats", menuName = "ScriptableObjects/UpgradeStats/SpeedUpgrade", order = 1)]
 public class SpeedUpgradeStats : UpgradeStats
 {
-    public int extraSpeed;
+    public float extraSpeed;
     
     public override void Apply(StatsManager manager)
     {
